dev: prevent writing to readonly config

+++ packages/cspell-config-lib/src/CSpellConfigFileReaderWriter.ts
@@ -3,7 +3,7 @@ import type { FileLoaderMiddleware } from './FileLoader.js';
 import type { IO } from './IO.js';
 import { getDeserializer, getLoader, getSerializer } from './middlewareHelper.js';
 import type { DeserializerNext, SerializerMiddleware } from './Serializer.js';
-import type { TextFile, TextFileRef } from './TextFile.js';
+import type { TextFileRef } from './TextFile.js';
 import { toURL } from './util/toURL.js';
 
 export interface CSpellConfigFileReaderWriter {
@@ -16,7 +16,7 @@ export interface CSpellConfigFileReaderWriter {
 
 export class CSpellConfigFileReaderWriterImpl implements CSpellConfigFileReaderWriter {
     /**
-     * @param io - an optional injectable IO interface. The default it to use the file system.
+     * @param io - an optional injectable IO interface. The default is to use the file system.
      * @param deserializers - Additional deserializers to use when reading a config file. The order of the deserializers is
      *    important. The last one in the list will be the first one to be called.
      */
@@ -35,16 +35,13 @@ export class CSpellConfigFileReaderWriterImpl implements CSpellConfigFileReaderW
         return getDeserializer(this.middleware);
     }
 
-    deserialize(file: TextFile): CSpellConfigFile {
-        return this.getDeserializer()(file);
-    }
-
     serialize(configFile: ICSpellConfigFile): string {
         const serializer = getSerializer(this.middleware);
         return serializer(configFile);
     }
 
     async writeConfig(configFile: ICSpellConfigFile): Promise<TextFileRef> {
+        if (configFile.readonly) throw new Error(`Config file is readonly: ${configFile.url.href}`);
         const content = this.serialize(configFile);
         await this.io.writeFile({ url: configFile.url, content });
         return { url: configFile.url };
